pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pykit"
version = "0.0.0"
description = "A logging framework for FRC robots running python"
readme = "README.md"
authors = [{ name = "Luke Maxwell", email = "luke@whsrobotics.org" }]
license = "MIT"
keywords = ["pykit"]
dependencies = [
    "pyntcore >=2025.1.1",
    "robotpy-wpimath >=2025.1.1",
    "robotpy-commands-v2 >=2025.1.1",
    "robotpy-wpiutil >=2025.1.1",
    "robotpy-hal >=2025.1.1"
]
requires-python = ">=3.10"

[project.urls]
Homepage = "https://github.com/1757WestwoodRobotics/pykit"
